chore: Update GDAL installation in CD workflow and remove unnecessary flag
Update the GDAL installation in the CD workflow and remove an unnecessary flag. This ensures that the correct version of GDAL (3.8.4) is installed and used for the project. Additionally, the virtual environment creation step has been removed as it is no longer needed.

Update the conda build configuration file to include Python versions 3.9, 3.10, and 3.11, while removing version 3.8.